3-structured-output/pyproject.toml
[project]
name = "3-structured-output"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-generativeai>=0.8.5",
    "pypdf>=6.0.0",
    "requests>=2.32.5",
]
